requirements.txt
pyro-ppl==1.8.1
torch==1.11.0
numpy==1.22.3